+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### To be Released
 
+* fix(domains): replace DomainsUpdate with Domain\*Certificate [#731](https://github.com/Scalingo/cli/pull/731)
 * chore(go): use go 1.17 [#728](https://github.com/Scalingo/cli/pull/728)
 * feat(app): Region detection from Git remote [#724](https://github.com/Scalingo/cli/pull/724)
 * feat(app): add a flag --force to destroy an app without interactive confirmation [#721](https://github.com/Scalingo/cli/pull/721)