+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### To be Released
 
+### 1.23.0
+
 * fix(domains): replace DomainsUpdate with Domain\*Certificate [#731](https://github.com/Scalingo/cli/pull/731)
 * chore(go): use go 1.17 [#728](https://github.com/Scalingo/cli/pull/728)
 * feat(app): Region detection from Git remote [#724](https://github.com/Scalingo/cli/pull/724)